requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==2.0.1
packaging==24.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
sqlparse==0.5.1
threadpoolctl==3.5.0
uritemplate==4.1.1
urllib3==2.2.2
